requirements.txt
openai
streamlit
python-dotenv
requests
gunicorn
flask
